+++ Runtime/Menus/AbstractMenu.cs
@@ -10,38 +10,17 @@ namespace ActionCode.UI
     [RequireComponent(typeof(GraphicRaycaster))]
     public abstract class AbstractMenu : CanvasViewer
     {
-        [SerializeField, Tooltip("The local GraphicRaycaster component.")]
-        protected GraphicRaycaster raycaster;
-
-        /// <summary>
-        /// Whether the Menu is visible.
-        /// </summary>
-        public override bool Visible
-        {
-            get => base.Visible;
-            set
-            {
-                base.Visible = value;
-                raycaster.enabled = value;
-            }
-        }
-
-        protected override void Reset()
-        {
-            base.Reset();
-            raycaster = GetComponent<GraphicRaycaster>();
-        }
-
-        protected virtual void Start()
-        {
-            if (Visible) Show();
-        }
-
+        protected virtual void Start() => CheckVisibility();
         protected virtual void OnEnable() => BindButtonsEvents();
         protected virtual void OnDisable() => UnBindButtonsEvents();
 
         protected abstract void BindButtonsEvents();
         protected abstract void UnBindButtonsEvents();
+
+        private void CheckVisibility()
+        {
+            if (Visible) Show();
+        }
 
         [ContextMenu("Create a Traditional Menu")]
         private void CreateTraditionalMenu()