+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - DelayedButton waits for seconds in realtime
 - Remove AudiSource and SelectableGroups from Popup component
 - Unsealed Popup component
+- CanvasViewer enables/disables GameObject when Visible changes
 
 ### Removed
 - MenuEventTrigger component